pyproject.toml
[project]
name = "rag-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.10",
    "langgraph-prebuilt>=1.0.13",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
